Cleaning up the interface
