+++ rocq-skylabs-cpp2v/src/PrintDecl.cpp
@@ -6,7 +6,6 @@
 #include "Assert.hpp"
 #include "ClangPrinter.hpp"
 #include "CoqPrinter.hpp"
-#include "DefaultTemplateAlias.hpp"
 #include "DeclVisitorWithArgs.h"
 #include "Formatter.hpp"
 #include "Logging.hpp"
@@ -304,214 +303,7 @@ static std::string getTemplateParamName(const NamedDecl &decl) {
     return "__template_param";
 }
 
-static fmt::Formatter &printId(CoqPrinter &print, StringRef name) {
-    guard::ctor _{print, "Nid", false};
-    return print.str(name);
-}
-
-static fmt::Formatter &printTemplateBaseName(CoqPrinter &print,
-                                             const NamedDecl &decl,
-                                             ClangPrinter &cprint) {
-    auto ctx = decl.getDeclContext();
-    while (ctx && !ctx->isTranslationUnit() && !isa<NamedDecl>(ctx))
-        ctx = ctx->getParent();
-
-    if (!ctx || ctx->isTranslationUnit()) {
-        guard::ctor _{print, "Nglobal", false};
-        return printId(print, decl.getName());
-    }
-
-    guard::ctor _{print, "Nscoped", false};
-    cprint.printName(print, *cast<NamedDecl>(Decl::castFromDeclContext(ctx)))
-        << fmt::nbsp;
-    return printId(print, decl.getName());
-}
-
-static fmt::Formatter &printTemplateParam(CoqPrinter &print,
-                                          const NamedDecl *param,
-                                          ClangPrinter &cprint) {
-    return cprint.printTemplateParam(print, param, loc::of(param));
-}
-
-static fmt::Formatter &printTemplateParamArg(CoqPrinter &print,
-                                             const NamedDecl *param,
-                                             ClangPrinter &cprint) {
-    return cprint.printTemplateArg(print, param, loc::of(param));
-}
-
-static fmt::Formatter &printTemplateNameWithArgs(
-    CoqPrinter &print, const NamedDecl &decl, ClangPrinter &cprint,
-    llvm::function_ref<void()> print_args) {
-    guard::ctor _{print, "Ninst", false};
-    printTemplateBaseName(print, decl, cprint) << fmt::nbsp;
-    print_args();
-    return print.output();
-}
-
-static fmt::Formatter &
-printDefaultAliasKey(CoqPrinter &print, const NamedDecl &decl,
-                     ArrayRef<const NamedDecl *> params, unsigned keep,
-                     ClangPrinter &cprint) {
-    return printTemplateNameWithArgs(print, decl, cprint, [&]() {
-        print.list(params.take_front(keep), [&](const NamedDecl *param) {
-            printTemplateParamArg(print, param, cprint);
-        });
-    });
-}
-
-static bool collectTemplateParamsForDecl(const Decl &decl,
-                                         std::vector<const NamedDecl *> &out) {
-    if (auto *ctx = decl.getDeclContext()) {
-        if (!ctx->isTranslationUnit()) {
-            if (!collectTemplateParamsForDecl(*Decl::castFromDeclContext(ctx),
-                                              out))
-                return false;
-        }
-    }
-
-    const TemplateParameterList *params = nullptr;
-    if (auto *record = dyn_cast<CXXRecordDecl>(&decl)) {
-        if (auto *templ = record->getDescribedClassTemplate())
-            params = templ->getTemplateParameters();
-    } else if (auto *alias = dyn_cast<TypeAliasDecl>(&decl)) {
-        if (auto *templ = alias->getDescribedAliasTemplate())
-            params = templ->getTemplateParameters();
-    }
-
-    if (!params)
-        return true;
-
-    for (auto *param : params->asArray()) {
-        if (!isa<TemplateTypeParmDecl>(param) &&
-            !isa<NonTypeTemplateParmDecl>(param))
-            return false;
-        if (param->isParameterPack())
-            return false;
-        out.push_back(param);
-    }
-
-    return true;
-}
-
-static bool hasDuplicateTemplateParamNames(
-    ArrayRef<const NamedDecl *> first, ArrayRef<const NamedDecl *> second) {
-    std::vector<std::string> names;
-    names.reserve(first.size() + second.size());
-
-    auto add_name = [&](const NamedDecl *param) {
-        auto name = getTemplateParamName(*param);
-        if (llvm::is_contained(names, name))
-            return false;
-        names.push_back(std::move(name));
-        return true;
-    };
-
-    for (auto *param : first)
-        if (!add_name(param))
-            return true;
-    for (auto *param : second)
-        if (!add_name(param))
-            return true;
-
-    return false;
-}
-
 } // namespace
-
-void printDefaultTemplateAliases(const Decl *decl, CoqPrinter &print,
-                                 ClangPrinter &cprint) {
-    if (!print.templates() || !cprint.printTypedefs())
-        return;
-
-    const NamedDecl *templated_decl = nullptr;
-    const TemplateParameterList *template_params = nullptr;
-
-    if (auto *record = dyn_cast_or_null<CXXRecordDecl>(decl)) {
-        if (auto *templ = record->getDescribedClassTemplate()) {
-            templated_decl = record;
-            template_params = templ->getTemplateParameters();
-        }
-    } else if (auto *alias = dyn_cast_or_null<TypeAliasDecl>(decl)) {
-        if (auto *templ = alias->getDescribedAliasTemplate()) {
-            templated_decl = alias;
-            template_params = templ->getTemplateParameters();
-        }
-    }
-
-    if (!templated_decl || !template_params)
-        return;
-
-    auto params = template_params->asArray();
-    if (params.empty())
-        return;
-
-    unsigned first_default = params.size();
-    std::vector<const NamedDecl *> prefix_params;
-    prefix_params.reserve(params.size());
-
-    for (unsigned i = 0; i < params.size(); ++i) {
-        if (params[i]->isParameterPack())
-            return;
-
-        if (auto *param = dyn_cast<TemplateTypeParmDecl>(params[i])) {
-            if (param->hasDefaultArgument()) {
-                first_default = std::min(first_default, i);
-            } else if (first_default != params.size()) {
-                return;
-            }
-        } else if (auto *param = dyn_cast<NonTypeTemplateParmDecl>(params[i])) {
-            if (param->hasDefaultArgument()) {
-                if (!default_template_alias::isSupportedValueDefault(param))
-                    return;
-                first_default = std::min(first_default, i);
-            } else if (first_default != params.size()) {
-                return;
-            }
-        } else {
-            return;
-        }
-        prefix_params.push_back(params[i]);
-    }
-
-    if (first_default == params.size())
-        return;
-
-    auto cp = cprint.withDecl(templated_decl);
-    std::vector<const NamedDecl *> enclosing_params;
-    if (auto *ctx = templated_decl->getDeclContext()) {
-        if (!ctx->isTranslationUnit())
-            if (!collectTemplateParamsForDecl(*Decl::castFromDeclContext(ctx),
-                                              enclosing_params))
-                return;
-    }
-    if (hasDuplicateTemplateParamNames(enclosing_params, prefix_params))
-        return;
-
-    for (unsigned keep = first_default; keep < params.size(); ++keep) {
-        {
-            guard::ctor _{print, "Dtemplated_typedef"};
-            {
-                guard::list _{print};
-                for (auto *param : enclosing_params)
-                    printTemplateParam(print, param, cp) << fmt::cons;
-                for (auto *param :
-                     ArrayRef<const NamedDecl *>(prefix_params).take_front(keep))
-                    printTemplateParam(print, param, cp) << fmt::cons;
-            }
-            print.output() << fmt::nbsp;
-            printDefaultAliasKey(print, *templated_decl, params, keep, cp)
-                << fmt::nbsp;
-            {
-                guard::ctor _{print, "Tnamed", false};
-                printTemplateNameWithArgs(print, *templated_decl, cp, [&]() {
-                    default_template_alias::printTargetArgs(print, params, keep,
-                                                            cp);
-                });
-            }
-        }
-        print.cons();
-    }
-}
 
 // Structures and unions
 namespace {
